refactor: simplify chat creation logic and fix package naming for exceptions
