+++ chat-service/src/main/java/com/example/chat_service/exceptions/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.chat_service.expections;
+package com.example.chat_service.exceptions;
 
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
