+++ chat-service/src/main/java/com/example/chat_service/exceptions/UnauthorizedOperationException.java
@@ -1,4 +1,4 @@
-package com.example.chat_service.expections;
+package com.example.chat_service.exceptions;
 
 public class UnauthorizedOperationException extends RuntimeException {
